requirements.txt
asgiref==3.7.1
autopep8==2.0.2
Django==3.2.19
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
flake8==5.0.4
importlib-metadata==4.2.0
mccabe==0.7.0
pep8==1.7.1
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2
zipp==3.15.0
